reorganize contents in marsutils
